+++ src/frontend/packages/lib_video/src/types/libs/altamoon-robust-websocket/index.d.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import 'altamoon-robust-websocket';
 
 declare module 'altamoon-robust-websocket' {