+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - lti:
   - position modal above tooltip
 - mutualize all CronJob in a CronJobList
+- make all packages shakeable
 
 ### Changed